+++ src/modules/itemsCounter.js
